organize server files
